+++ src/Providers/MonkyzServiceProvider.php
@@ -50,8 +50,5 @@ class MonkyzServiceProvider extends ServiceProvider
 		$this->commands([
 			\Lab1353\Monkyz\Commands\MonkyzTables::class
 		]);
-
-		// dependency
-		$this->app->register('Barryvdh\Elfinder\ElfinderServiceProvider');
 	}
 }